requirements.txt
Django==4.2.3
lightning==2.0.5
numpy==1.23.5
pandas==1.5.3
Pillow==10.0.0
torch==2.0.1+cu118
torchmetrics==1.0.0
tqdm==4.64.1
transformers==4.19.2
